requirements.txt
wheel
tqdm
lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy==1.0.3
imageio
git+https://github.com/rmurai0610/diff-gaussian-rasterization-w-pose.git
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
scikit-video
opencv-python
